drivers:adc:adc_demo: Use submit function instead of read_dev
Use submit function instead of read_dev for buffer readings.
Move adc_submit_samples and adc_demo_trigger_handler
implementations to iio_adc_demo.c.
